requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
Django==4.1
djantic==0.7.0
fastapi==0.79.0
h11==0.13.0
idna==3.3
Pillow==9.2.0
pydantic==1.9.1
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
PyJWT==2.4.0